requirements.txt
Django>=4.1.2,<4.2.0
djangorestframework>=3.14.0,<3.15.0
psycopg2>=2.9.4,<2.10